pyproject.toml
[project]
name = "malackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash",
    "pandas",
    "dash-bootstrap-components",
    "plotly>=6.3.1",
    "gunicorn",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
